# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: CUENTY API - Enterprise
  description: |
    API REST completa para el sistema CUENTY de gestión de suscripciones para el mercado mexicano.
    
    ## Características
    - Autenticación JWT + API Keys
    - Rate limiting: 100 requests/minuto
    - Webhooks para integraciones
    - SDKs disponibles: Node.js, Python, PHP
    
    ## Ambiente
    - Base URL Producción: https://api.cuenty.com/v1
    - Base URL Sandbox: https://sandbox.api.cuenty.com/v1
    
  version: 3.0.0
  contact:
    name: CUENTY API Support
    email: api@cuenty.com
    url: https://cuenty.com/docs
  license:
    name: Propietario
    url: https://cuenty.com/legal/api-terms
servers:
  - url: https://api.cuenty.com/v1
    description: Servidor de Producción
  - url: https://sandbox.api.cuenty.com/v1
    description: Servidor de Sandbox
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenido del endpoint /auth/login
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para desarrolladores externos
  
  # ==========================================================================
  # SCHEMAS
  # ==========================================================================
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_REQUEST
            message:
              type: string
              example: El campo email es requerido
            details:
              type: object
    
    # Auth
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: usuario@example.com
        password:
          type: string
          format: password
          example: password123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        user:
          $ref: '#/components/schemas/User'
        expiresIn:
          type: string
          example: 24h
    
    # User
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: usuario@example.com
        name:
          type: string
          example: Juan Pérez
        phone:
          type: string
          example: +525512345678
        role:
          type: string
          enum: [customer, admin]
          example: customer
        createdAt:
          type: string
          format: date-time
    
    # Subscription
...
    Webhook:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
          example: https://mi-app.com/webhooks/cuenty
        events:
          type: array
          items:
            type: string
          example: [payment.completed, subscription.activated]
        secret:
          type: string
          description: Secret para verificar firma HMAC
          example: whsec_123456789

# ==========================================================================
# PATHS
# ==========================================================================
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Iniciar sesión
      description: Autenticar usuario y obtener token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Registrar nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
                - phone
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                name:
                  type: string
                phone:
                  type: string
      responses:
        '201':
          description: Usuario creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/2fa/send:
    post:
      tags:
        - Authentication
      summary: Enviar código 2FA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [sms, whatsapp]
      responses:
        '200':
          description: Código enviado exitosamente
  /auth/2fa/verify:
    post:
      tags:
        - Authentication
      summary: Verificar código 2FA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: Código verificado exitosamente
        '400':
          description: Código inválido
  
  # ==========================================================================
  # USERS
  # ==========================================================================
...
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: Listar webhooks configurados
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Lista de webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
    
    post:
      tags:
        - Webhooks
      summary: Crear nuevo webhook
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Webhook'
      responses:
        '201':
          description: Webhook creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []